app/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
blinker==1.6.2
boto3==1.28.38
botocore==1.31.38
charset-normalizer==3.2.0
click==8.1.7
colorlog==6.7.0
exceptiongroup==1.1.3
fastapi==0.103.0
Flask==2.3.3
frozenlist==1.4.0
h11==0.14.0
hikari==2.0.0.dev120
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
multidict==6.0.4
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
